Add eval & add input methods to MacVim, apply to file open
Now possible to evaluate expressions in a Vim process from MacVim [MMBackend
evaluateExpression:], and to send arbitrary input to Vim [MMVimController
addVimInput:].  Expression evaluation is used to query Vim state from within
MacVim.

When opening a file use these methods to check if it is already loaded, if so
raise the corresponding window (doesn't work for multiple files).