Window and view refactoring
Window and view code was refactored to better accomodate the transition
to the ATSUI text rendering code.  View specific code has been
completely moved into MMVimView and window code is now all in
MMWindowController; the window controller is also 'full-screen aware'.
Finally, the text storage is no longer referenced outside MMTextView --
the idea is that the text view is the public class for handling the
work that MMTextStorage used to be the public interface for.

This refactoring also brings some improvements:
- window resizing is more responsive
- full-screen window memory leak fixed
- proper fix for erroneous 'buffer modified' warning
- full-screen no longer tried to set 'nil-title' (this caused backend
  connection problems)
- changing lines/columns in full-screen works properly
- open dialog now works in full-screen on Tiger