Find Next/Previous and Use Selection for Find menus
Implemented "Find Next", "Find Previous" and "Use Selection for Find"
menus.  In the process the Edit->Find menu has been restructured.  To
accomodate these changes the code changed so that the Find Pasteboard is
updated also when :set @/ = ... is called (previously the Find
Pasteboard was only updated when an actual search was executed).