Fix vert split problems in Core Text renderer
The display is no longer corrupted on :vsp when using the Core Text
renderer.  Also fixes a problem where background would not be cleared
correctly when left scrollbar was visible.

Remove a hack in live resizing code when using CT.  This may cause the
window to flicker more during live resize (have to find a better fix).