Backend initiates window zooming
Don't change the window size immediately upon pressing the zoom button.
Instead send the message to Vim and let it resize before the window
does.  This avoids flickering problem when zooming with the Core Text
renderer.

The zooming is also slightly smarter about which window size to restore
if the toolbar was hidden while the window was zoomed.