Allow window resizing across multiple screens
It is now possible to drag to resize a window to be larger than one
screen.  However, if the window size changes programmatically (due to
scrollbars or toolbar hiding/showing, etc.) then the window will be
constrained to fit on the screen holding most of the window.  This is
perhaps annoying but at least it is consistent with the way Terminal.app
works (try opening a new tab when a Terminal window spans two screens).