Pass arguments to Vim on command line
Do not pass arguments via DO since some options only work if the files
to open were passed on the command line (e.g. "%" flag in 'viminfo').
Also, if a file is encrypted and opened from Spotlight the Spotlight
search text is passed as input and it ends up being used as password for
the encryption key, so the file opens up encrypted.

The above problems still persist if Quickstart is enabled.