Do not open files in "unused windows"
A window is "unused" if it has just been opened with Cmd+n.  Previously
a file opened from Finder would be opened in an unused window if one was
available.  This leads to problems with some Vim options, e.g. trying to
open an encrypted file from Spotlight will fail.  The code that has been
removed also had some quite ugly hacks and it is not entirely clear
whether the reuse of unused windows was entirely intuitive.

(The "Open files from applications" option is not affected by this
change.)