Add +balloon_eval support
The tracking rect code had to be pulled to get tool tips to work.  This
has at least one negative drawback (the mouse cursor does not change
back to an arrow when moved outside the window) and there may be more.

Changes to 'bdlay' are only detected after disabling and re-enabling
'beval'.