Handle resolution changes while in full screen
If the screen resolution changes (either due to an external monitor
being unplugged or because the settings were changed manually) the full
screen window is now resized to match the new resolution.  Before this
commit full screen was exited on such an event.

This fixes a bug where the full screen option got out of sync after the
resolution changed.