Force autorelease pool flush after an FSEvent
The top-level autorelease pool only gets cleared out on application
events, not including things like the FSEvent handler.  If MacVim is in
the background while ~/.vim changes, autoreleased objects will carry on
building up until the users switches back to it, resulting in a high
memory water mark and a lot of needless deallocs.  See
http://www.mikeash.com/pyblog/more-fun-with-autorelease.html