Default to using CoreText renderer
The default renderer is too "loose" on OS X 10.8 whereas the CoreText
renderer looks the same across all platforms.  Furthermore, the CoreText
renderer is many times faster.

A downside to the CoreText renderer is that it does not support the text
input protocols properly, so e.g. three-finger tapping to show
dictionary results does not work.  It also does not show marked text as
it is being drawn if the "inline marked text" option is disabled.