Use “inoremap” for <M-BS> and <D-BS>
As a Mac user I’m used to using Option-Delete to delete to the beginning of the previous word; MacVim sets up this behavior with its gvimrc. But I also like to use Control-W to delete to the beginning of the line. This change makes it possible for these two mappings to coexist without a problem.