Fix font fallback with CoreText renderer
`CTFontCreateForString()` only consults the cascade list embedded in the
font, with fallback to the global default cascade list if the font has
none. We don't have cascade lists in our fonts, so this means that it
will never return a font that's not in the global cascade list. This is
unfortunate because it causes .LastResort glyphs to be shown for
characters for which the user does have a font installed.

The simplest workaround for this seems to be to create a CTLineRef for
the string and then check its runs for the fonts chosen. CoreText uses
more advanced fallback behavior in this scenario, still giving
preference to the global cascade list but it will find other installed
fonts instead of ending up with .LastResort.

Also, fix a memory leak where we weren't releasing the `newFontRef` if
the call to `CTFontGetGlyphsForCharacters()` returned false.

Fixes #243.