Fix flashing during window resizing.
The root cause of flashing/flickering was that on each resize event,
AppKit called drawRect: on the MMCoreTextView, expecting it to fill an
empty, newly-sized buffer with content. It has nothing to draw, so the
view would show up as black until the BatchDrawMsgID reply arrived from
the backend and triggered another draw. (The MMCoreTextView was
display:ed twice for each resize).

This change does three things:

1. Implements -[MMCoreTextView setFrameSize:] to begin an
   NSAnimationContext grouping before calling super, which postpones
   display:/drawRect: until the grouping ends, and ends the grouping in
   performBatchDrawWithData:.

2. Makes *all* resize messages sent to the backend synchronous, with a
   1s timeout. This seems bad, but actually helps avoid spamming the
   backend with resize events, since the app only generates resize
   events as fast as we handle them (i.e. if we take time to process one
   resize, then the next resize event will have the current size of the
   window, potentially skipping a bunch in the middle that the backend
   never would have had time to handle).

3. Gets rid of some hacks resolved by 1 and 2, like MKVimController
   delaying the call to -[MMWindowController presentWindow:].