Really turn off ligatures when not turned on
According to the documentation for kCTLigatureAttributeName [1], the
value 0 turns on essential ligatures for the rendering of the given
text. It seems that some monospace fonts with ligature support, eg
Iosevka and Fira Code, declares all ligatures as essential. The
rendering is "a bit" broken when you use one of such fonts and turn off
ligatures. When we do not include kCTLigatureAttributeName at all in the
attributes dictionary, it seems that no ligatures are rendered, which is
what we want.

[1] https://developer.apple.com/documentation/coretext/kctligatureattributename