Refactor Ctrl-6 fix to be clearer and only trigger under Japanese IME
Only limiting this condition to be under Japanese IME helps prevents
unintended effects on other IMEs as well, as this bug is only known to
happen under Japanese IME with Windows shortcuts.

Also update comments to be clear that most of the time you do want to go
through the interpretKeyEvents: path.