Fix scrollbars in Mojave dark mode not rendering properly
Dark mode scrollbars's background are rendered in a translucent color so
that it would overlay on top of the background nearly, even though we
are using legacy scrollbars with a dedicated space. This is unlike light
mode scrollbars which render with a concrete color. This means if the
background has uncleared rendering it would result in some oddities.
Just fix it by making sure we first render the scroll track with the
current background color before drawing the system overlay.

Known issues:
- The scroll knob will look quite light and hard to see if the
  background color is bright. Consider this OK for now as users who use
  dark mode will likely have a dark-ish background color in Vim.
- If both vertical and horizontal scrollbars are enabled, the corners
  will look black or sometimes filled with rendering artifacts. Will fix
  this later. One way to fix is to always fill the Vim view with a
  background color but that seems to slow things down a little bit
  because setNeedsDisplay seems to be called too much without a rect, so
  need to fix that first.
- Vertical scrollbars' positions are sometimes set incorrectly. That's a
  separate bug and will be addressed later.
- MacVim currently doesn't support overlay scrollbars which have been
  introduced in macOS since 10.7. Investigate that option too.