Expose "automatically install updates" option in preference pane
Currently the updater (Sparkle) supports an option to automatically
download and install updates in the background. The only user-visible
way to turn it on/off is by checking the checkbox in the new update
dialog box but that dialog box only shows up if auto-update is off,
meaning that there's no way to undo setting that setting. Fix this by
exposing the setting in user preference panel.

Also, previously auto-update would even still keep on working even if
"check for updates" is off, since it has precedence. Make the preference
pane un-check auto-update if "check for updates" is turned off.

Fix #808