Fix `Class MMBackend` duplication bug on launch
This issue was detected on both Homebrew and Github released MacVim for
v8.1 on a fresh install of OS High Sierra with no .vim or .vimrc configuration

Basically, when starting vim from the command line, it produces the
following error:

```
objc[4513]: Class MMBackend is implemented in both /usr/local/Cellar/macvim/8.1-153/MacVim.app/Contents/bin/../MacOS/Vim (0x10b303f38) and /usr/local/Cellar/macvim/8.1-153/MacVim.app/Contents/MacOS/Vim (0x217f5bf38). One of the two will be used. Which one is undefined.
```

If launching graphical interface, the error shows up twice.

The issue was traced to the use of '/../' in the binary path for
launching the application which somehow pollutes the library path for
the executable, finding copies of the MMBackend class in both the
absolute path and the quazi-relative path.

To resolve this, the relative component was removed from the path
resolution of the actual Vim binary by actually dropping down a
directory from the script location.  It also simplifies the
dereferencing of any link chains to that script.